Next.js part 2 (pages to app)
