+++ .eslintrc.json
@@ -1,10 +1,22 @@
 {
+  "root": true,
   "extends": [
     "next/core-web-vitals",
+    "prettier",
+    "plugin:jsx-a11y/recommended",
     "eslint:recommended",
+    "plugin:react/recommended",
+    "plugin:react-hooks/recommended",
     "plugin:@typescript-eslint/recommended"
   ],
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
-  "root": true
+  "plugins": ["simple-import-sort", "@typescript-eslint"],
+  "rules": {
+    "jsx-a11y/anchor-is-valid": "error",
+    "simple-import-sort/imports": "error",
+    "simple-import-sort/exports": "error",
+    "import/first": "error",
+    "import/newline-after-import": "error",
+    "import/no-duplicates": "error"
+  }
 }
