+++ package.json
@@ -28,6 +28,7 @@
     "sharp": "^0.34.1"
   },
   "devDependencies": {
+    "@eslint/js": "^9.24.0",
     "@tailwindcss/postcss": "^4.1.3",
     "@types/node": "^22.14.0",
     "@types/react": "^19.1.0",
@@ -43,31 +44,7 @@
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "prettier": "^3.5.3",
     "tailwindcss": "^4.1.3",
-    "typescript": "^5.8.3"
-  },
-  "eslintConfig": {
-    "root": true,
-    "extends": [
-      "next/core-web-vitals",
-      "prettier",
-      "plugin:jsx-a11y/recommended",
-      "eslint:recommended",
-      "plugin:react/recommended",
-      "plugin:react-hooks/recommended",
-      "plugin:@typescript-eslint/recommended"
-    ],
-    "parser": "@typescript-eslint/parser",
-    "plugins": [
-      "simple-import-sort",
-      "@typescript-eslint"
-    ],
-    "rules": {
-      "jsx-a11y/anchor-is-valid": "error",
-      "simple-import-sort/imports": "error",
-      "simple-import-sort/exports": "error",
-      "import/first": "error",
-      "import/newline-after-import": "error",
-      "import/no-duplicates": "error"
-    }
+    "typescript": "^5.8.3",
+    "typescript-eslint": "^8.29.1"
   }
 }
