+++ src/app/not-found.tsx
@@ -1,6 +1,8 @@
-import Header from '@/components/Header';
-import Footer from '@/components/Footer';
 import Head from 'next/head';
+import React from 'react';
+
+import Footer from '@/components/Footer';
+import Header from '@/components/Header';
 
 const _404 = () => (
   <>
