+++ src/app/sitemap.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
+<url><loc>https://www.utdnebula.com/robots.txt</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/manifest.json</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/sitemap.xml</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/apple-icon.png</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/opengraph-image.png</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/projects/api</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/membership</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/projects/trends</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/about/governance/2022-2023</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/about/governance/2023-2024</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/about/mission</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/contact</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/projects/skedge</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/projects/planner</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/projects/jupiter</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/about/governance</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+<url><loc>https://www.utdnebula.com/resources/calendar</loc><lastmod>2025-04-10T22:13:13.808Z</lastmod><changefreq>daily</changefreq><priority>0.7</priority></url>
+</urlset>
