+++ src/components/Footer.tsx
@@ -1,26 +1,22 @@
 import Image from 'next/image';
 import Link from 'next/link';
-import LogoName from '@/../public/logo-name-white.svg';
-import Arrow from '@/../public/icons/arrow-white.svg';
-import Discord from '@/../public/icons/join-discord-white.svg';
-import Instagram from '@/../public/icons/instagram-white.svg';
-import LinkedIn from '@/../public/icons/linkedin-white.svg';
+import React from 'react';
+
 import GitHub from '@/../public/icons/github-white.svg';
+import Instagram from '@/../public/icons/instagram-white.svg';
+import Discord from '@/../public/icons/join-discord-white.svg';
+import LinkedIn from '@/../public/icons/linkedin-white.svg';
+import LogoName from '@/../public/logo-name-white.svg';
+import ScrollUpButton from '@/components/ScrollUpButton';
 
 const linkClasses = 'underline decoration-transparent hover:decoration-inherit transition';
 
-const Footer = () => {
+export default function Footer() {
   return (
     <footer className="bg-royal text-white w-full rounded-t-[3.125rem] pb-10 md:px-40 px-8 text-sm mt-20">
       <div className="flex justify-between py-16 sm:py-28 gap-8">
         <Image src={LogoName} alt="logo" width="360" height="53" className="shrink min-w-0" />
-        <button
-          onClick={() => window.scrollTo(0, 0)}
-          className="items-center flex flex-col rounded-full p-2 transition border border-white/0 hover:border-white"
-        >
-          <Image src={Arrow} alt="arrow" width="20" height="20" className="rotate-180" />
-          Top
-        </button>
+        <ScrollUpButton />
       </div>
       <div className="grid grid-cols-2 lg:grid-cols-4 gap-10">
         <div className="flex flex-col items-start gap-3">
@@ -68,6 +64,7 @@ const Footer = () => {
             className={linkClasses}
             href="https://nebula-labs.atlassian.net/wiki/spaces/ND/overview?homepageId=23822536"
             target="_blank"
+            rel="noreferrer"
           >
             Design Guide
           </a>
@@ -75,6 +72,7 @@ const Footer = () => {
             className={linkClasses}
             href="https://nebula-labs.atlassian.net/wiki/x/0YD5AQ"
             target="_blank"
+            rel="noreferrer"
           >
             Confluence
           </a>
@@ -84,6 +82,7 @@ const Footer = () => {
             className="mb-6 hover:scale-105 transition"
             href="https://discord.utdnebula.com/"
             target="_blank"
+            rel="noreferrer"
           >
             <Image src={Discord} alt="discord" width="200" height="60" />
           </a>
@@ -91,6 +90,7 @@ const Footer = () => {
             className={linkClasses + ' flex items-center gap-2'}
             href="https://instagram.com/utdnebula"
             target="_blank"
+            rel="noreferrer"
           >
             <Image src={Instagram} alt="Instagram logo" width="30" height="30" />
             Instagram
@@ -99,6 +99,7 @@ const Footer = () => {
             className={linkClasses + ' flex items-center gap-2'}
             href="https://linkedin.com/company/utdnebula"
             target="_blank"
+            rel="noreferrer"
           >
             <Image src={LinkedIn} alt="LinkedIn logo" width="30" height="30" />
             LinkedIn
@@ -107,6 +108,7 @@ const Footer = () => {
             className={linkClasses + ' flex items-center gap-2'}
             href="https://github.com/utdnebula"
             target="_blank"
+            rel="noreferrer"
           >
             <Image src={GitHub} alt="GitHub logo" width="30" height="30" />
             GitHub
@@ -132,6 +134,4 @@ const Footer = () => {
       </div>
     </footer>
   );
-};
-
-export default Footer;
+}
