+++ src/components/Governance.tsx
@@ -1,18 +1,20 @@
-import { useState, useEffect } from 'react';
-import Header from '@/components/Header';
-import Footer from '@/components/Footer';
-import PeriodLinks from '@/components/PeriodLinks';
-import Image from 'next/image';
-import { PopulatedGoveranceGroup } from '@/lib/period-populator';
-import { Person } from '@/data/person-dictionary';
+'use client';
 
-import OrgMatrix from '@/../public/org-matrix.png';
-import LinkedIn from '@/../public/icons/linkedin-royal.svg';
+import Image from 'next/image';
+import React, { useEffect, useState } from 'react';
+
 import Email from '@/../public/icons/email.svg';
+import LinkedIn from '@/../public/icons/linkedin-royal.svg';
+import OrgMatrix from '@/../public/org-matrix.png';
+import Footer from '@/components/Footer';
+import Header from '@/components/Header';
+import PeriodLinks from '@/components/PeriodLinks';
+import { Person } from '@/data/person-dictionary';
+import { PopulatedGoveranceGroup } from '@/lib/period-populator';
 
 const fallbackSrc = '/governance/blank.jpg';
 
-const LeadershipCard = (props: Person) => {
+function LeadershipCard(props: Person) {
   const [src, setSrc] = useState(`/governance/${props.netId}.jpg`);
   useEffect(() => {
     setSrc(`/governance/${props.netId}.jpg`);
@@ -26,8 +28,8 @@ const LeadershipCard = (props: Person) => {
         width={280}
         className="w-full aspect-square rounded-3xl"
         key={props.netId}
-        onLoadingComplete={(result) => {
-          if (result.naturalWidth === 0) {
+        onLoad={(result) => {
+          if (result.currentTarget.naturalWidth === 0) {
             // Broken image
             setSrc(fallbackSrc);
           }
@@ -56,9 +58,9 @@ const LeadershipCard = (props: Person) => {
       </div>
     </div>
   );
-};
+}
 
-const LeadershipGroup = (props: PopulatedGoveranceGroup) => {
+function LeadershipGroup(props: PopulatedGoveranceGroup) {
   return (
     <div className="px-8 lg:px-16 xl:px-32 py-24 flex flex-col items-center gap-12">
       <h2 className="text-5xl font-bold text-center">{props.name}</h2>
@@ -70,7 +72,7 @@ const LeadershipGroup = (props: PopulatedGoveranceGroup) => {
       </div>
     </div>
   );
-};
+}
 
 interface GovernanceProps {
   period: string;
@@ -82,27 +84,27 @@ interface GovernanceProps {
   };
 }
 
-const Governance = (props: GovernanceProps) => (
-  <>
-    <Header text={'Our ' + (props.isCurrent ? '' : props.period + ' ') + 'Leadership Team'} />
-    {props.data.map((group) => (
-      <LeadershipGroup {...group} key={group.name} />
-    ))}
-    <div className="px-8 lg:px-16 xl:px-32 py-24 flex flex-col items-center gap-12">
-      <h2 className="text-5xl font-bold text-center">Organizational Structure</h2>
-      <p className="text-3xl">
-        Nebula Labs is based on a matrix style organizational structure. This means each member has
-        two points of contact: their project lead and division head. Project leads are overseen by
-        the Executive Director and division heads are overseen by the Vice President.
-      </p>
-      <Image
-        src={OrgMatrix}
-        alt="Organization matrix chart showing the overlap of projects and divisions"
-      />
-    </div>
-    <PeriodLinks name="Historical governance periods" {...props.periodLinks} />
-    <Footer />
-  </>
-);
-
-export default Governance;
+export default function Governance(props: GovernanceProps) {
+  return (
+    <>
+      <Header text={'Our ' + (props.isCurrent ? '' : props.period + ' ') + 'Leadership Team'} />
+      {props.data.map((group) => (
+        <LeadershipGroup {...group} key={group.name} />
+      ))}
+      <div className="px-8 lg:px-16 xl:px-32 py-24 flex flex-col items-center gap-12">
+        <h2 className="text-5xl font-bold text-center">Organizational Structure</h2>
+        <p className="text-3xl">
+          Nebula Labs is based on a matrix style organizational structure. This means each member
+          has two points of contact: their project lead and division head. Project leads are
+          overseen by the Executive Director and division heads are overseen by the Vice President.
+        </p>
+        <Image
+          src={OrgMatrix}
+          alt="Organization matrix chart showing the overlap of projects and divisions"
+        />
+      </div>
+      <PeriodLinks name="Historical governance periods" {...props.periodLinks} />
+      <Footer />
+    </>
+  );
+}
